Q4 considerations
Both Logistic Regression (LR) and Gaussian Discriminant Analysis (GDA) are used for classification and give a slight different decision boundaries. GDA makes an assumption about the probability distribution of the p(x|y). And it can be easily proved that for GDA, if the initial assumptions about distribution of p(x|y) (obeys to Gaussian distribution) and p(y)(obeys to Bernoulli distribution) are true, then p(y|x) can be expressed as sigmoid. But vice-versa may not true. It means that GDA makes more specific assumptions about the dataset than LR and if those assumptions are true then it works better than LR.
So when the dataset obeys or roughly obeys the normal distribution(Gaussian distribution), GDA can fit it well, because GDA uses more information to build the model, and the stronger modelling assuptions is helpful.
But when the dataset does not obey a normal distribution, then logistic regression is more effective because it makes fewer assumptions, and the model is stronger and more robust, so the stronger modelling assuptions is harmful.
Another advantage of generative learning model (GDA) is that it can build a strong model using less data than discriminative learning models(LR).